Fix certificate generation on server start
Fix vcxproj files generation
Fixes after review
Link fix
Refactor checkable models
VMS-2293: Create dialog for access rights setup

<==== Mercurial data ====>
Branch: prod_3.0.0

+++ client/src/ui/models/audit/audit_log_model.cpp
@@ -666,7 +666,7 @@ bool QnAuditLogModel::setData(const QModelIndex &index, const QVariant &value, i
 
     if (role == Qt::CheckStateRole)
     {
-        if (value == Qt::Checked)
+        if (value.toInt() == Qt::Checked)
             m_index->at(index.row())->addParam(CheckedParamName, "1");
         else
             m_index->at(index.row())->removeParam(CheckedParamName);
